Address more PR feedback: move WhatIf dir up, split SDK client, revert unrelated changes
- Move Implementation/DeploymentStacks/DeploymentStacksWhatIf -> Implementation/DeploymentStacksWhatIf (same level as DeploymentStacks per Dante's comment)

- Create separate DeploymentStacksWhatIfSdkClient for WhatIf operations (distinct resource type) to make review cleaner and prevent regressions in regular stacks SDK client

- Revert unintended changes to ExportAzureResourceGroupCmdlet.cs, NewResourceManagerSdkClient.cs, NewResourcesExtensions.cs, PSScriptStatus.cs, ErrorResponse.cs (generated SDK)

- Remove dead Track 2 SDK code (ArmClient, Azure.Core usings, AzureContextCredential) that was no longer needed

- Update all cmdlets to use the new DeploymentStacksWhatIfSdkClient property on base class
