+++ src/Resources/ResourceManager/Implementation/DeploymentStacks/NewAzSubscriptionDeploymentStack.cs
@@ -1,4 +1,4 @@
-﻿// ----------------------------------------------------------------------------------
+// ----------------------------------------------------------------------------------
 //
 // Copyright Microsoft Corporation
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -101,7 +101,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                 // Handle WhatIf scenario
                 if (MyInvocation.BoundParameters.ContainsKey("WhatIf") && ((SwitchParameter)MyInvocation.BoundParameters["WhatIf"]).ToBool())
                 {
-                    var whatIfResult = DeploymentStacksSdkClient.ExecuteSubscriptionDeploymentStackWhatIf(
+                    var whatIfResult = DeploymentStacksWhatIfSdkClient.ExecuteSubscriptionDeploymentStackWhatIf(
                         deploymentStackName: Name,
                         location: Location,
                         templateFile: TemplateFile,
