+++ src/Resources/ResourceManager/Implementation/DeploymentStacksWhatIf/NewAzResourceGroupDeploymentStackWhatIf.cs
