+++ src/Resources/ResourceManager/Implementation/DeploymentStacksWhatIf/SetAzManagementGroupDeploymentStackWhatIf.cs
