+++ src/Resources/ResourceManager/Implementation/DeploymentStacksWhatIf/SetAzResourceGroupDeploymentStackWhatIf.cs
