+++ src/Resources/ResourceManager/ResourceManager.csproj
@@ -1,48 +1,44 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-	<PropertyGroup>
-		<PsModuleName>Resources</PsModuleName>
-	</PropertyGroup>
+  <PropertyGroup>
+    <PsModuleName>Resources</PsModuleName>
+  </PropertyGroup>
 
-	<Import Project="$(MSBuildThisFileDirectory)..\..\Az.props" />
+  <Import Project="$(MSBuildThisFileDirectory)..\..\Az.props" />
 
-	<PropertyGroup>
-		<AssemblyName>$(AzAssemblyPrefix)ResourceManager</AssemblyName>
-		<RootNamespace>$(LegacyAssemblyPrefix)ResourceManager.Cmdlets</RootNamespace>
-	</PropertyGroup>
+  <PropertyGroup>
+    <AssemblyName>$(AzAssemblyPrefix)ResourceManager</AssemblyName>
+    <RootNamespace>$(LegacyAssemblyPrefix)ResourceManager.Cmdlets</RootNamespace>
+  </PropertyGroup>
 
-	<ItemGroup>
-		<PackageReference Include="Microsoft.Azure.Management.ResourceManager" Version="3.17.3-preview" />
-		<PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="2.2.0" />
-		<!-- Update Azure.Core to support the new DeploymentStacks SDK (imported from Az.props) -->
-		<PackageReference Update="Azure.Core" Version="1.51.1" />
-		<!-- New Azure SDK for Deployment Stacks What-If operations -->
-		<PackageReference Include="Azure.ResourceManager.Resources.DeploymentStacks" Version="1.0.0" />
-	</ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Azure.Management.ResourceManager" Version="3.17.3-preview" />
+    <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="2.2.0" />
+  </ItemGroup>
 
-	<ItemGroup>
-		<ProjectReference Include="..\Resources.Management.Sdk\Resources.Management.Sdk.csproj" />
-	</ItemGroup>
+  <ItemGroup>
+	<ProjectReference Include="..\Resources.Management.Sdk\Resources.Management.Sdk.csproj" />
+  </ItemGroup>
 
-	<ItemGroup>
-		<Compile Update="Properties\Resources.Designer.cs">
-			<DesignTime>True</DesignTime>
-			<AutoGen>True</AutoGen>
-			<DependentUpon>Resources.resx</DependentUpon>
-		</Compile>
-	</ItemGroup>
+  <ItemGroup>
+    <Compile Update="Properties\Resources.Designer.cs">
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>Resources.resx</DependentUpon>
+    </Compile>
+  </ItemGroup>
 
-	<ItemGroup>
-		<EmbeddedResource Update="Properties\Resources.resx">
-			<Generator>ResXFileCodeGenerator</Generator>
-			<LastGenOutput>Resources.Designer.cs</LastGenOutput>
-		</EmbeddedResource>
-	</ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Update="Properties\Resources.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+  </ItemGroup>
 
-	<ItemGroup>
-		<None Update="ResourceManagerStartup.ps1" CopyToOutputDirectory="Never" />
-	</ItemGroup>
+  <ItemGroup>
+    <None Update="ResourceManagerStartup.ps1" CopyToOutputDirectory="Never" />
+  </ItemGroup>
 
-	<Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory).., build.proj))\src\Az.Post.props" />
+  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory).., build.proj))\src\Az.Post.props" />
 
 </Project>
