+++ src/Resources/ResourceManager/SdkClient/DeploymentStacksSdkClient.cs
@@ -37,9 +37,6 @@ using Newtonsoft.Json;
 using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.DeploymentStacks;
 using Microsoft.Azure.Commands.ResourceManager.Cmdlets.Json;
 using Microsoft.WindowsAzure.Commands.Common;
-using Azure.Core;
-using Azure.ResourceManager;
-using Azure.ResourceManager.Resources;
 
 namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
 {
@@ -87,28 +84,6 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             }
 
             set { this.resourceManagerSdkClient = value; }
-        }
-
-        /// <summary>
-        /// Field that holds the ARM client instance for Track 2 SDK
-        /// </summary>
-        private ArmClient armClient;
-
-        /// <summary>
-        /// Gets the ARM client for Track 2 SDK operations (e.g., What-If)
-        /// </summary>
-        private ArmClient ArmClient
-        {
-            get
-            {
-                if (this.armClient == null && this.azureContext != null)
-                {
-                    var credential = new AzureContextCredential(this.azureContext);
-                    var armClientOptions = new ArmClientOptions();
-                    this.armClient = new ArmClient(credential, this.azureContext.Subscription.Id, armClientOptions);
-                }
-                return this.armClient;
-            }
         }
 
         private enum DeploymentStackScope
@@ -176,7 +151,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                     }
                     else
                     {
-                        throw new PSArgumentException(erex.Body.Error.Message);
+                        throw new PSArgumentException(erex.Body.Message);
                     }
                 }
 
@@ -360,7 +335,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                     }
                     else
                     {
-                        throw new PSArgumentException(erex.Body.Error.Message);
+                        throw new PSArgumentException(erex.Body.Message);
                     }
                 }
 
@@ -439,7 +414,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                     }
                     else
                     {
-                        throw new PSArgumentException(erex.Body.Error.Message);
+                        throw new PSArgumentException(erex.Body.Message);
                     }
                 }
 
@@ -1246,951 +1221,6 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             }
 
             return convertedErrors;
-        }
-
-        /// <summary>
-        /// Converts Azure SDK What-If result to PowerShell model
-        /// </summary>
-        private PSDeploymentStackWhatIfResult ConvertToPSDeploymentStackWhatIfResult(object sdkWhatIfResult)
-        {
-            // Serialize SDK result to JSON and deserialize to PS model
-            // This provides a flexible conversion that handles the complex nested structure
-            var json = JsonConvert.SerializeObject(sdkWhatIfResult);
-            var psResult = JsonConvert.DeserializeObject<PSDeploymentStackWhatIfResult>(json);
-            return psResult;
-        }
-
-        /// <summary>
-        /// Creates a DeploymentStacksWhatIfResult request object for the What-If operation
-        /// </summary>
-        private DeploymentStacksWhatIfResult CreateDeploymentStacksWhatIfResult(
-            string location,
-            string templateFile,
-            string templateUri,
-            string templateSpec,
-            Hashtable templateObject,
-            string parameterUri,
-            Hashtable parameters,
-            string description,
-            string resourcesCleanupAction,
-            string resourceGroupsCleanupAction,
-            string managementGroupsCleanupAction,
-            string deploymentScope,
-            string denySettingsMode,
-            string[] denySettingsExcludedPrincipals,
-            string[] denySettingsExcludedActions,
-            bool denySettingsApplyToChildScopes,
-            string deploymentStackName,
-            string resourceGroupName,
-            string managementGroupId,
-            string stackResourceId = null,
-            string retentionInterval = null,
-            string validationLevel = null,
-            string debugSettingDetailLevel = null,
-            bool bypassStackOutOfSyncError = false)
-        {
-            // Convert PS models to SDK models
-            var actionOnUnmanage = new AzSdkModels.ActionOnUnmanage(
-                resources: resourcesCleanupAction,
-                resourceGroups: resourceGroupsCleanupAction,
-                managementGroups: managementGroupsCleanupAction);
-
-            var denySettings = new AzSdkModels.DenySettings(
-                mode: denySettingsMode,
-                excludedPrincipals: denySettingsExcludedPrincipals?.ToList(),
-                excludedActions: denySettingsExcludedActions?.ToList(),
-                applyToChildScopes: denySettingsApplyToChildScopes);
-
-            // Use user-provided stackResourceId if available, else construct from name + scope
-            string resolvedStackResourceId = stackResourceId;
-            if (string.IsNullOrEmpty(resolvedStackResourceId))
-            {
-                if (!string.IsNullOrEmpty(resourceGroupName))
-                {
-                    resolvedStackResourceId = $"/subscriptions/{azureContext.Subscription.Id}/resourceGroups/{resourceGroupName}/providers/Microsoft.Resources/deploymentStacks/{deploymentStackName}";
-                }
-                else if (!string.IsNullOrEmpty(managementGroupId))
-                {
-                    resolvedStackResourceId = $"/providers/Microsoft.Management/managementGroups/{managementGroupId}/providers/Microsoft.Resources/deploymentStacks/{deploymentStackName}";
-                }
-                else
-                {
-                    resolvedStackResourceId = $"/subscriptions/{azureContext.Subscription.Id}/providers/Microsoft.Resources/deploymentStacks/{deploymentStackName}";
-                }
-            }
-
-            var properties = new DeploymentStacksWhatIfResultProperties(
-                actionOnUnmanage: actionOnUnmanage,
-                denySettings: denySettings,
-                deploymentStackResourceId: resolvedStackResourceId,
-                retentionInterval: !string.IsNullOrEmpty(retentionInterval) ? System.Xml.XmlConvert.ToTimeSpan(retentionInterval) : TimeSpan.FromHours(2))
-            {
-                Description = description,
-                DeploymentScope = deploymentScope,
-                ValidationLevel = validationLevel,
-                DebugSetting = !string.IsNullOrEmpty(debugSettingDetailLevel)
-                    ? new AzSdkModels.DeploymentStacksDebugSetting(debugSettingDetailLevel)
-                    : null
-            };
-
-            // Evaluate Template
-            if (templateSpec != null)
-            {
-                properties.TemplateLink = new DeploymentStacksTemplateLink
-                {
-                    Id = templateSpec,
-                };
-            }
-            else if (Uri.IsWellFormedUriString(templateUri, UriKind.Absolute))
-            {
-                properties.TemplateLink = new DeploymentStacksTemplateLink
-                {
-                    Uri = templateUri,
-                };
-            }
-            else if (!string.IsNullOrEmpty(templateFile))
-            {
-                var templateJObject = FileUtilities.DataStore.ReadFileAsStream(templateFile).FromJson<JObject>();
-                properties.Template = templateJObject.ToObject<Dictionary<string, object>>();
-            }
-            else
-            {
-                var templateJToken = templateObject.ToJToken();
-                properties.Template = templateJToken.ToObject<Dictionary<string, object>>();
-            }
-
-            // Evaluate Template Parameters
-            if (Uri.IsWellFormedUriString(parameterUri, UriKind.Absolute))
-            {
-                properties.ParametersLink = new DeploymentStacksParametersLink
-                {
-                    Uri = parameterUri
-                };
-            }
-            else if (parameters != null)
-            {
-                properties.Parameters = ConvertParameterHashtableToDictionary(parameters);
-            }
-
-            var whatIfResult = new DeploymentStacksWhatIfResult
-            {
-                Location = location,
-                Properties = properties
-            };
-
-            return whatIfResult;
-        }
-
-        #region What-If Resource CRUD Operations
-
-        public PSDeploymentStackWhatIfResult GetResourceGroupDeploymentStackWhatIfResult(string resourceGroupName, string stackName, bool throwIfNotExists = true)
-        {
-            try
-            {
-                var result = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.Get(resourceGroupName, stackName);
-                return ConvertToPSDeploymentStackWhatIfResult(result);
-            }
-            catch (Exception ex)
-            {
-                if (ex is DeploymentStacksErrorException dex)
-                {
-                    if (dex.Response.StatusCode == System.Net.HttpStatusCode.NotFound)
-                    {
-                        if (!throwIfNotExists)
-                            return null;
-                        throw new PSArgumentException($"WhatIf result '{stackName}' in Resource Group '{resourceGroupName}' not found.");
-                    }
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                }
-                throw;
-            }
-        }
-
-        public PSDeploymentStackWhatIfResult GetResourceGroupDeploymentStackWhatIfResultWithPropertyChanges(string resourceGroupName, string stackName)
-        {
-            try
-            {
-                var result = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.WhatIf(resourceGroupName, stackName);
-                return ConvertToPSDeploymentStackWhatIfResult(result);
-            }
-            catch (Exception ex)
-            {
-                if (ex is DeploymentStacksErrorException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public IList<PSDeploymentStackWhatIfResult> ListResourceGroupDeploymentStackWhatIfResults(string resourceGroupName)
-        {
-            try
-            {
-                var list = new List<PSDeploymentStackWhatIfResult>();
-                var results = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.List(resourceGroupName);
-                list.AddRange(results.Select(r => ConvertToPSDeploymentStackWhatIfResult(r)));
-                while (results.NextPageLink != null)
-                {
-                    results = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.ListNext(results.NextPageLink);
-                    list.AddRange(results.Select(r => ConvertToPSDeploymentStackWhatIfResult(r)));
-                }
-                return list;
-            }
-            catch (Exception ex)
-            {
-                if (ex is DeploymentStacksErrorException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public void DeleteResourceGroupDeploymentStackWhatIfResult(string resourceGroupName, string stackName)
-        {
-            try
-            {
-                DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.Delete(resourceGroupName, stackName);
-            }
-            catch (Exception ex)
-            {
-                if (ex is DeploymentStacksErrorException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public PSDeploymentStackWhatIfResult GetSubscriptionDeploymentStackWhatIfResult(string stackName, bool throwIfNotExists = true)
-        {
-            try
-            {
-                var result = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.Get(stackName);
-                return ConvertToPSDeploymentStackWhatIfResult(result);
-            }
-            catch (Exception ex)
-            {
-                if (ex is DeploymentStacksErrorException dex)
-                {
-                    if (dex.Response.StatusCode == System.Net.HttpStatusCode.NotFound)
-                    {
-                        if (!throwIfNotExists)
-                            return null;
-                        throw new PSArgumentException($"WhatIf result '{stackName}' not found in the current subscription.");
-                    }
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                }
-                throw;
-            }
-        }
-
-        public PSDeploymentStackWhatIfResult GetSubscriptionDeploymentStackWhatIfResultWithPropertyChanges(string stackName)
-        {
-            try
-            {
-                var result = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.WhatIf(stackName);
-                return ConvertToPSDeploymentStackWhatIfResult(result);
-            }
-            catch (Exception ex)
-            {
-                if (ex is DeploymentStacksErrorException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public IList<PSDeploymentStackWhatIfResult> ListSubscriptionDeploymentStackWhatIfResults()
-        {
-            try
-            {
-                var list = new List<PSDeploymentStackWhatIfResult>();
-                var results = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.List();
-                list.AddRange(results.Select(r => ConvertToPSDeploymentStackWhatIfResult(r)));
-                while (results.NextPageLink != null)
-                {
-                    results = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.ListNext(results.NextPageLink);
-                    list.AddRange(results.Select(r => ConvertToPSDeploymentStackWhatIfResult(r)));
-                }
-                return list;
-            }
-            catch (Exception ex)
-            {
-                if (ex is DeploymentStacksErrorException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public void DeleteSubscriptionDeploymentStackWhatIfResult(string stackName)
-        {
-            try
-            {
-                DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.Delete(stackName);
-            }
-            catch (Exception ex)
-            {
-                if (ex is DeploymentStacksErrorException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public PSDeploymentStackWhatIfResult GetManagementGroupDeploymentStackWhatIfResult(string managementGroupId, string stackName, bool throwIfNotExists = true)
-        {
-            try
-            {
-                var result = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.Get(managementGroupId, stackName);
-                return ConvertToPSDeploymentStackWhatIfResult(result);
-            }
-            catch (Exception ex)
-            {
-                if (ex is DeploymentStacksErrorException dex)
-                {
-                    if (dex.Response.StatusCode == System.Net.HttpStatusCode.NotFound)
-                    {
-                        if (!throwIfNotExists)
-                            return null;
-                        throw new PSArgumentException($"WhatIf result '{stackName}' in Management Group '{managementGroupId}' not found.");
-                    }
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                }
-                throw;
-            }
-        }
-
-        public PSDeploymentStackWhatIfResult GetManagementGroupDeploymentStackWhatIfResultWithPropertyChanges(string managementGroupId, string stackName)
-        {
-            try
-            {
-                var result = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.WhatIf(managementGroupId, stackName);
-                return ConvertToPSDeploymentStackWhatIfResult(result);
-            }
-            catch (Exception ex)
-            {
-                if (ex is DeploymentStacksErrorException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public IList<PSDeploymentStackWhatIfResult> ListManagementGroupDeploymentStackWhatIfResults(string managementGroupId)
-        {
-            try
-            {
-                var list = new List<PSDeploymentStackWhatIfResult>();
-                var results = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.List(managementGroupId);
-                list.AddRange(results.Select(r => ConvertToPSDeploymentStackWhatIfResult(r)));
-                while (results.NextPageLink != null)
-                {
-                    results = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.ListNext(results.NextPageLink);
-                    list.AddRange(results.Select(r => ConvertToPSDeploymentStackWhatIfResult(r)));
-                }
-                return list;
-            }
-            catch (Exception ex)
-            {
-                if (ex is DeploymentStacksErrorException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        public void DeleteManagementGroupDeploymentStackWhatIfResult(string managementGroupId, string stackName)
-        {
-            try
-            {
-                DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.Delete(managementGroupId, stackName);
-            }
-            catch (Exception ex)
-            {
-                if (ex is DeploymentStacksErrorException dex)
-                    throw new PSArgumentException(dex.Body?.Error?.Message ?? dex.Message);
-                throw;
-            }
-        }
-
-        #endregion
-
-        #region What-If Operations
-
-        /// <summary>
-        /// Executes a what-if operation for a deployment stack.
-        /// Main entry point that determines scope and routes to appropriate method.
-        /// NOTE: What-If for Deployment Stacks is not yet supported by the Azure REST API.
-        /// </summary>
-        public PSDeploymentStackWhatIfResult ExecuteDeploymentStackWhatIf(PSDeploymentStackWhatIfParameters parameters)
-        {
-            if (parameters == null)
-            {
-                throw new ArgumentNullException(nameof(parameters));
-            }
-
-            // Determine scope and call appropriate method
-            if (!string.IsNullOrEmpty(parameters.ResourceGroupName))
-            {
-                return ExecuteResourceGroupDeploymentStackWhatIf(
-                    parameters.StackName,
-                    parameters.ResourceGroupName,
-                    parameters.TemplateFile,
-                    parameters.TemplateUri,
-                    parameters.TemplateSpecId,
-                    parameters.TemplateObject,
-                    parameters.TemplateParameterUri,
-                    parameters.TemplateParameterObject,
-                    parameters.Description,
-                    parameters.ResourcesCleanupAction,
-                    parameters.ResourceGroupsCleanupAction,
-                    parameters.ManagementGroupsCleanupAction,
-                    parameters.DenySettingsMode,
-                    parameters.DenySettingsExcludedPrincipals,
-                    parameters.DenySettingsExcludedActions,
-                    parameters.DenySettingsApplyToChildScopes,
-                    parameters.BypassStackOutOfSyncError,
-                    parameters.StackResourceId,
-                    parameters.RetentionInterval,
-                    parameters.ValidationLevel,
-                    parameters.DebugSettingDetailLevel);
-            }
-            else if (!string.IsNullOrEmpty(parameters.ManagementGroupId))
-            {
-                return ExecuteManagementGroupDeploymentStackWhatIf(
-                    parameters.StackName,
-                    parameters.ManagementGroupId,
-                    parameters.Location,
-                    parameters.TemplateFile,
-                    parameters.TemplateUri,
-                    parameters.TemplateSpecId,
-                    parameters.TemplateObject,
-                    parameters.TemplateParameterUri,
-                    parameters.TemplateParameterObject,
-                    parameters.Description,
-                    parameters.ResourcesCleanupAction,
-                    parameters.ResourceGroupsCleanupAction,
-                    parameters.ManagementGroupsCleanupAction,
-                    parameters.DeploymentScope,
-                    parameters.DenySettingsMode,
-                    parameters.DenySettingsExcludedPrincipals,
-                    parameters.DenySettingsExcludedActions,
-                    parameters.DenySettingsApplyToChildScopes,
-                    parameters.BypassStackOutOfSyncError,
-                    parameters.StackResourceId,
-                    parameters.RetentionInterval,
-                    parameters.ValidationLevel,
-                    parameters.DebugSettingDetailLevel);
-            }
-            else
-            {
-                return ExecuteSubscriptionDeploymentStackWhatIf(
-                    parameters.StackName,
-                    parameters.Location,
-                    parameters.TemplateFile,
-                    parameters.TemplateUri,
-                    parameters.TemplateSpecId,
-                    parameters.TemplateObject,
-                    parameters.TemplateParameterUri,
-                    parameters.TemplateParameterObject,
-                    parameters.Description,
-                    parameters.ResourcesCleanupAction,
-                    parameters.ResourceGroupsCleanupAction,
-                    parameters.ManagementGroupsCleanupAction,
-                    parameters.DeploymentScope,
-                    parameters.DenySettingsMode,
-                    parameters.DenySettingsExcludedPrincipals,
-                    parameters.DenySettingsExcludedActions,
-                    parameters.DenySettingsApplyToChildScopes,
-                    parameters.BypassStackOutOfSyncError,
-                    parameters.StackResourceId,
-                    parameters.RetentionInterval,
-                    parameters.ValidationLevel,
-                    parameters.DebugSettingDetailLevel);
-            }
-        }
-
-        /// <summary>
-        /// Executes a what-if operation for a deployment stack at resource group scope.
-        /// This uses the two-step process: CreateOrUpdate the WhatIf result, then call WhatIf to get changes.
-        /// </summary>
-        public PSDeploymentStackWhatIfResult ExecuteResourceGroupDeploymentStackWhatIf(
-            string deploymentStackName,
-            string resourceGroupName,
-            string templateFile,
-            string templateUri,
-            string templateSpec,
-            Hashtable templateObject,
-            string parameterUri,
-            Hashtable parameters,
-            string description,
-            string resourcesCleanupAction,
-            string resourceGroupsCleanupAction,
-            string managementGroupsCleanupAction,
-            string denySettingsMode,
-            string[] denySettingsExcludedPrincipals,
-            string[] denySettingsExcludedActions,
-            bool denySettingsApplyToChildScopes,
-            bool bypassStackOutOfSyncError,
-            string stackResourceId = null,
-            string retentionInterval = null,
-            string validationLevel = null,
-            string debugSettingDetailLevel = null)
-        {
-            try
-            {
-                WriteVerbose($"Executing What-If for deployment stack '{deploymentStackName}' in resource group '{resourceGroupName}'");
-
-                // Step 1: Create what-if result object to submit
-                var whatIfRequest = CreateDeploymentStacksWhatIfResult(
-                    location: null,
-                    templateFile,
-                    templateUri,
-                    templateSpec,
-                    templateObject,
-                    parameterUri,
-                    parameters,
-                    description,
-                    resourcesCleanupAction,
-                    resourceGroupsCleanupAction,
-                    managementGroupsCleanupAction,
-                    deploymentScope: null,
-                    denySettingsMode,
-                    denySettingsExcludedPrincipals,
-                    denySettingsExcludedActions,
-                    denySettingsApplyToChildScopes,
-                    deploymentStackName,
-                    resourceGroupName,
-                    null,
-                    stackResourceId,
-                    retentionInterval,
-                    validationLevel,
-                    debugSettingDetailLevel,
-                    bypassStackOutOfSyncError
-                );
-
-                // Step 2: Submit the WhatIf request (CreateOrUpdate creates/updates a WhatIf result resource)
-                WriteVerbose("Creating What-If result resource...");
-                var createResult = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.CreateOrUpdate(
-                    resourceGroupName,
-                    deploymentStackName,
-                    whatIfRequest);
-
-                // The createResult should already contain the changes - check if we need to poll or if it's complete
-                WriteVerbose($"Initial What-If result state: {createResult.Properties?.ProvisioningState}");
-                
-                // Poll for completion if the result is being computed asynchronously
-                DeploymentStacksWhatIfResult finalResult = createResult;
-                const int maxPollingAttempts = 60; // 5 minutes max
-                int pollingInterval = 5; // seconds
-                int attempts = 0;
-                
-                while (finalResult.Properties?.ProvisioningState == "Running" && attempts < maxPollingAttempts)
-                {
-                    WriteVerbose($"Polling What-If result... attempt {attempts + 1}");
-                    System.Threading.Thread.Sleep(pollingInterval * 1000);
-                    finalResult = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.Get(
-                        resourceGroupName,
-                        deploymentStackName);
-                    attempts++;
-                }
-
-                WriteVerbose($"Final What-If result state: {finalResult.Properties?.ProvisioningState}");
-
-                // Call WhatIf POST to get result with property changes (delta) populated
-                WriteVerbose("Fetching What-If result with property changes...");
-                try
-                {
-                    var resultWithChanges = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtResourceGroup.WhatIf(
-                        resourceGroupName,
-                        deploymentStackName);
-                    return ConvertToPSDeploymentStackWhatIfResult(resultWithChanges);
-                }
-                catch (Exception postEx)
-                {
-                    WriteVerbose($"WhatIf POST not available, returning GET result: {postEx.Message}");
-                    return ConvertToPSDeploymentStackWhatIfResult(finalResult);
-                }
-            }
-            catch (ErrorResponseException ex) when (ex.Response.StatusCode == System.Net.HttpStatusCode.NotFound || 
-                                                     (ex.Response.StatusCode == System.Net.HttpStatusCode.BadRequest && 
-                                                      ex.Body?.Error?.Message?.IndexOf("what-if result", StringComparison.OrdinalIgnoreCase) >= 0))
-            {
-                // What-If API not available - provide informational message
-                WriteWarning("Deployment Stacks What-If API is not yet available in this region or subscription.");
-                WriteWarning("The deployment stack would be created/updated with the following configuration:");
-                WriteWarning($"  Name: {deploymentStackName}");
-                WriteWarning($"  Resource Group: {resourceGroupName}");
-                WriteWarning($"  Action on unmanaged resources: {resourcesCleanupAction}");
-                WriteWarning($"  Action on unmanaged resource groups: {resourceGroupsCleanupAction}");
-                WriteWarning($"  Deny Settings Mode: {denySettingsMode}");
-                
-                // Return a basic result indicating What-If is not supported
-                var notSupportedResult = new PSDeploymentStackWhatIfResult
-                {
-                    Name = deploymentStackName,
-                    Type = "Microsoft.Resources/deploymentStacks",
-                    Properties = new DeploymentStackWhatIfProperties
-                    {
-                        ProvisioningState = "What-If API not available"
-                    }
-                };
-                
-                return notSupportedResult;
-            }
-            catch (ErrorResponseException ex)
-            {
-                WriteError($"Error executing What-If (HTTP {ex.Response?.StatusCode}): {ex.Body?.Error?.Message ?? ex.Message}");
-                if (ex.Body?.Error?.Details != null && ex.Body.Error.Details.Any())
-                {
-                    foreach (var detail in ex.Body.Error.Details)
-                    {
-                        WriteError($"  - {detail.Code}: {detail.Message}");
-                    }
-                }
-                throw;
-            }
-            catch (Exception ex)
-            {
-                WriteError($"Error executing What-If: {ex.Message}");
-                throw;
-            }
-        }
-
-        /// <summary>
-        /// Executes a what-if operation for a deployment stack at subscription scope.
-        /// </summary>
-        public PSDeploymentStackWhatIfResult ExecuteSubscriptionDeploymentStackWhatIf(
-            string deploymentStackName,
-            string location,
-            string templateFile,
-            string templateUri,
-            string templateSpec,
-            Hashtable templateObject,
-            string parameterUri,
-            Hashtable parameters,
-            string description,
-            string resourcesCleanupAction,
-            string resourceGroupsCleanupAction,
-            string managementGroupsCleanupAction,
-            string deploymentScope,
-            string denySettingsMode,
-            string[] denySettingsExcludedPrincipals,
-            string[] denySettingsExcludedActions,
-            bool denySettingsApplyToChildScopes,
-            bool bypassStackOutOfSyncError,
-            string stackResourceId = null,
-            string retentionInterval = null,
-            string validationLevel = null,
-            string debugSettingDetailLevel = null)
-        {
-            try
-            {
-                WriteVerbose($"Executing What-If for deployment stack '{deploymentStackName}' at subscription scope");
-
-                // Create what-if result object to submit
-                var whatIfRequest = CreateDeploymentStacksWhatIfResult(
-                    location,
-                    templateFile,
-                    templateUri,
-                    templateSpec,
-                    templateObject,
-                    parameterUri,
-                    parameters,
-                    description,
-                    resourcesCleanupAction,
-                    resourceGroupsCleanupAction,
-                    managementGroupsCleanupAction,
-                    deploymentScope,
-                    denySettingsMode,
-                    denySettingsExcludedPrincipals,
-                    denySettingsExcludedActions,
-                    denySettingsApplyToChildScopes,
-                    deploymentStackName,
-                    null,
-                    null,
-                    stackResourceId,
-                    retentionInterval,
-                    validationLevel,
-                    debugSettingDetailLevel,
-                    bypassStackOutOfSyncError
-                );
-
-                // Execute What-If operation using Track 1 SDK
-                WriteVerbose("Creating What-If result resource...");
-                var createResult = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.CreateOrUpdate(
-                    deploymentStackName,
-                    whatIfRequest);
-
-                // Poll for completion if needed
-                WriteVerbose($"Initial What-If result state: {createResult.Properties?.ProvisioningState}");
-                
-                DeploymentStacksWhatIfResult finalResult = createResult;
-                const int maxPollingAttempts = 60;
-                int pollingInterval = 5;
-                int attempts = 0;
-                
-                while (finalResult.Properties?.ProvisioningState == "Running" && attempts < maxPollingAttempts)
-                {
-                    WriteVerbose($"Polling What-If result... attempt {attempts + 1}");
-                    System.Threading.Thread.Sleep(pollingInterval * 1000);
-                    finalResult = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.Get(deploymentStackName);
-                    attempts++;
-                }
-
-                WriteVerbose($"Final What-If result state: {finalResult.Properties?.ProvisioningState}");
-
-                // Call WhatIf POST to get result with property changes (delta) populated
-                WriteVerbose("Fetching What-If result with property changes...");
-                try
-                {
-                    var resultWithChanges = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtSubscription.WhatIf(deploymentStackName);
-                    return ConvertToPSDeploymentStackWhatIfResult(resultWithChanges);
-                }
-                catch (Exception postEx)
-                {
-                    WriteVerbose($"WhatIf POST not available, returning GET result: {postEx.Message}");
-                    return ConvertToPSDeploymentStackWhatIfResult(finalResult);
-                }
-            }
-            catch (ErrorResponseException ex) when (ex.Response.StatusCode == System.Net.HttpStatusCode.NotFound || 
-                                                     (ex.Response.StatusCode == System.Net.HttpStatusCode.BadRequest && 
-                                                      ex.Body?.Error?.Message?.IndexOf("what-if result", StringComparison.OrdinalIgnoreCase) >= 0))
-            {
-                // What-If API not available - provide informational message
-                WriteWarning("Deployment Stacks What-If API is not yet available in this region or subscription.");
-                WriteWarning("The deployment stack would be created/updated with the following configuration:");
-                WriteWarning($"  Name: {deploymentStackName}");
-                WriteWarning($"  Location: {location}");
-                WriteWarning($"  Action on unmanaged resources: {resourcesCleanupAction}");
-                WriteWarning($"  Action on unmanaged resource groups: {resourceGroupsCleanupAction}");
-                WriteWarning($"  Deny Settings Mode: {denySettingsMode}");
-                
-                // Return a basic result indicating What-If is not supported
-                var notSupportedResult = new PSDeploymentStackWhatIfResult
-                {
-                    Name = deploymentStackName,
-                    Type = "Microsoft.Resources/deploymentStacks",
-                    Properties = new DeploymentStackWhatIfProperties
-                    {
-                        ProvisioningState = "What-If API not available"
-                    }
-                };
-                
-                return notSupportedResult;
-            }
-            catch (ErrorResponseException ex)
-            {
-                WriteError($"Error executing What-If (HTTP {ex.Response?.StatusCode}): {ex.Body?.Error?.Message ?? ex.Message}");
-                if (ex.Body?.Error?.Details != null && ex.Body.Error.Details.Any())
-                {
-                    foreach (var detail in ex.Body.Error.Details)
-                    {
-                        WriteError($"  - {detail.Code}: {detail.Message}");
-                    }
-                }
-                throw;
-            }
-            catch (Exception ex)
-            {
-                WriteError($"Error executing What-If: {ex.Message}");
-                throw;
-            }
-        }
-
-        /// <summary>
-        /// Executes a what-if operation for a deployment stack at management group scope.
-        /// </summary>
-        public PSDeploymentStackWhatIfResult ExecuteManagementGroupDeploymentStackWhatIf(
-            string deploymentStackName,
-            string managementGroupId,
-            string location,
-            string templateFile,
-            string templateUri,
-            string templateSpec,
-            Hashtable templateObject,
-            string parameterUri,
-            Hashtable parameters,
-            string description,
-            string resourcesCleanupAction,
-            string resourceGroupsCleanupAction,
-            string managementGroupsCleanupAction,
-            string deploymentScope,
-            string denySettingsMode,
-            string[] denySettingsExcludedPrincipals,
-            string[] denySettingsExcludedActions,
-            bool denySettingsApplyToChildScopes,
-            bool bypassStackOutOfSyncError,
-            string stackResourceId = null,
-            string retentionInterval = null,
-            string validationLevel = null,
-            string debugSettingDetailLevel = null)
-        {
-            try
-            {
-                WriteVerbose($"Executing What-If for deployment stack '{deploymentStackName}' in management group '{managementGroupId}'");
-
-                // Create what-if result object to submit
-                var whatIfRequest = CreateDeploymentStacksWhatIfResult(
-                    location,
-                    templateFile,
-                    templateUri,
-                    templateSpec,
-                    templateObject,
-                    parameterUri,
-                    parameters,
-                    description,
-                    resourcesCleanupAction,
-                    resourceGroupsCleanupAction,
-                    managementGroupsCleanupAction,
-                    deploymentScope,
-                    denySettingsMode,
-                    denySettingsExcludedPrincipals,
-                    denySettingsExcludedActions,
-                    denySettingsApplyToChildScopes,
-                    deploymentStackName,
-                    null,
-                    managementGroupId,
-                    stackResourceId,
-                    retentionInterval,
-                    validationLevel,
-                    debugSettingDetailLevel,
-                    bypassStackOutOfSyncError
-                );
-
-                // Execute What-If operation using Track 1 SDK
-                WriteVerbose("Creating What-If result resource...");
-                var createResult = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.CreateOrUpdate(
-                    managementGroupId,
-                    deploymentStackName,
-                    whatIfRequest);
-
-                // Poll for completion if needed
-                WriteVerbose($"Initial What-If result state: {createResult.Properties?.ProvisioningState}");
-                
-                DeploymentStacksWhatIfResult finalResult = createResult;
-                const int maxPollingAttempts = 60;
-                int pollingInterval = 5;
-                int attempts = 0;
-                
-                while (finalResult.Properties?.ProvisioningState == "Running" && attempts < maxPollingAttempts)
-                {
-                    WriteVerbose($"Polling What-If result... attempt {attempts + 1}");
-                    System.Threading.Thread.Sleep(pollingInterval * 1000);
-                    finalResult = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.Get(
-                        managementGroupId,
-                        deploymentStackName);
-                    attempts++;
-                }
-
-                WriteVerbose($"Final What-If result state: {finalResult.Properties?.ProvisioningState}");
-
-                // Call WhatIf POST to get result with property changes (delta) populated
-                WriteVerbose("Fetching What-If result with property changes...");
-                try
-                {
-                    var resultWithChanges = DeploymentStacksClient.DeploymentStacksWhatIfResultsAtManagementGroup.WhatIf(
-                        managementGroupId,
-                        deploymentStackName);
-                    return ConvertToPSDeploymentStackWhatIfResult(resultWithChanges);
-                }
-                catch (Exception postEx)
-                {
-                    WriteVerbose($"WhatIf POST not available, returning GET result: {postEx.Message}");
-                    return ConvertToPSDeploymentStackWhatIfResult(finalResult);
-                }
-            }
-            catch (ErrorResponseException ex) when (ex.Response.StatusCode == System.Net.HttpStatusCode.NotFound || 
-                                                     (ex.Response.StatusCode == System.Net.HttpStatusCode.BadRequest && 
-                                                      ex.Body?.Error?.Message?.IndexOf("what-if result", StringComparison.OrdinalIgnoreCase) >= 0))
-            {
-                // What-If API not available - provide informational message
-                WriteWarning("Deployment Stacks What-If API is not yet available in this region or subscription.");
-                WriteWarning("The deployment stack would be created/updated with the following configuration:");
-                WriteWarning($"  Name: {deploymentStackName}");
-                WriteWarning($"  Management Group: {managementGroupId}");
-                WriteWarning($"  Location: {location}");
-                WriteWarning($"  Action on unmanaged resources: {resourcesCleanupAction}");
-                WriteWarning($"  Action on unmanaged resource groups: {resourceGroupsCleanupAction}");
-                WriteWarning($"  Deny Settings Mode: {denySettingsMode}");
-                
-                // Return a basic result indicating What-If is not supported
-                var notSupportedResult = new PSDeploymentStackWhatIfResult
-                {
-                    Name = deploymentStackName,
-                    Type = "Microsoft.Resources/deploymentStacks",
-                    Properties = new DeploymentStackWhatIfProperties
-                    {
-                        ProvisioningState = "What-If API not available"
-                    }
-                };
-                
-                return notSupportedResult;
-            }
-            catch (ErrorResponseException ex)
-            {
-                WriteError($"Error executing What-If (HTTP {ex.Response?.StatusCode}): {ex.Body?.Error?.Message ?? ex.Message}");
-                if (ex.Body?.Error?.Details != null && ex.Body.Error.Details.Any())
-                {
-                    foreach (var detail in ex.Body.Error.Details)
-                    {
-                        WriteError($"  - {detail.Code}: {detail.Message}");
-                    }
-                }
-                throw;
-            }
-            catch (Exception ex)
-            {
-                WriteError($"Error executing What-If: {ex.Message}");
-                throw;
-            }
-        }
-
-        #endregion
-    }
-
-    /// <summary>
-    /// Token credential implementation for Azure Context
-    /// </summary>
-    internal class AzureContextCredential : TokenCredential
-    {
-        private readonly IAzureContext context;
-
-        public AzureContextCredential(IAzureContext context)
-        {
-            this.context = context ?? throw new ArgumentNullException(nameof(context));
-        }
-
-        public override AccessToken GetToken(TokenRequestContext requestContext, CancellationToken cancellationToken)
-        {
-            var accessToken = AzureSession.Instance.AuthenticationFactory.Authenticate(
-                context.Account,
-                context.Environment,
-                context.Tenant.Id,
-                null,
-                ShowDialog.Never,
-                null,
-                requestContext.Scopes?.FirstOrDefault());
-
-            // Try to get expiration time from the token
-            DateTimeOffset expiresOn = DateTimeOffset.UtcNow.AddHours(1); // Default to 1 hour from now
-            
-            // Try to get ExpiresOn from extended properties
-            if (accessToken.ExtendedProperties != null && accessToken.ExtendedProperties.TryGetValue("ExpiresOn", out string expiresOnStr))
-            {
-                if (DateTimeOffset.TryParse(expiresOnStr, out var parsedDate))
-                {
-                    expiresOn = parsedDate;
-                }
-            }
-            else
-            {
-                // Try to use reflection to get ExpiresOn property if it exists
-                var expiresOnProperty = accessToken.GetType().GetProperty("ExpiresOn");
-                if (expiresOnProperty != null && expiresOnProperty.PropertyType == typeof(DateTimeOffset))
-                {
-                    expiresOn = (DateTimeOffset)expiresOnProperty.GetValue(accessToken);
-                }
-            }
-
-            return new AccessToken(accessToken.AccessToken, expiresOn);
-        }
-
-        public override ValueTask<AccessToken> GetTokenAsync(TokenRequestContext requestContext, CancellationToken cancellationToken)
-        {
-            return new ValueTask<AccessToken>(GetToken(requestContext, cancellationToken));
         }
     }
 }
