+++ src/Resources/Resources/Resources.csproj
@@ -18,8 +18,7 @@
 	<ItemGroup>
 		<PackageReference Include="Microsoft.Azure.Management.ManagementGroups" Version="2.0.0-beta" />
 		<PackageReference Include="Microsoft.Azure.Management.ResourceManager" Version="3.17.3-preview" />
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="10.0.2" />
-		<PackageReference Update="Azure.Core" Version="1.51.1" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.2" />
 	</ItemGroup>
 
 	<ItemGroup>
